+++ kroxylicious-kubernetes/kroxylicious-operator/src/test/java/io/kroxylicious/kubernetes/operator/reconciler/kafkaservice/StrimziKafkaSecondaryToKafkaServicePrimaryMapperTest.java
@@ -12,13 +12,14 @@ import org.junit.jupiter.api.Test;
 
 import io.javaoperatorsdk.operator.api.reconciler.EventSourceContext;
 import io.javaoperatorsdk.operator.processing.event.ResourceID;
+import io.strimzi.api.kafka.model.kafka.KafkaBuilder;
 
 import io.kroxylicious.kubernetes.api.v1alpha1.KafkaService;
+import io.kroxylicious.kubernetes.api.v1alpha1.KafkaServiceBuilder;
 
 import static io.kroxylicious.kubernetes.operator.reconciler.kafkaservice.MapperTestSupport.KAFKA;
 import static io.kroxylicious.kubernetes.operator.reconciler.kafkaservice.MapperTestSupport.SERVICE;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
 
 class StrimziKafkaSecondaryToKafkaServicePrimaryMapperTest {
 
@@ -36,18 +37,31 @@ class StrimziKafkaSecondaryToKafkaServicePrimaryMapperTest {
     }
 
     @Test
-    void shouldReturnIdsWhenApiServerUnavailable() {
+    void canMapFromNamespacedStrimziKafkaToKafkaServiceInDifferentNamespace() {
         // Given
-        EventSourceContext<KafkaService> context = mock();
-        MapperTestSupport.stubFailingListOperationClient(context);
-        MapperTestSupport.stubPrimaryCache(context, SERVICE);
-
+        KafkaService service = new KafkaServiceBuilder(SERVICE)
+                .editMetadata()
+                .withNamespace("my-proxy")
+                .endMetadata()
+                .editSpec()
+                .editStrimziKafkaRef()
+                .withNamespace("my-kafka")
+                .endStrimziKafkaRef()
+                .endSpec()
+                .build();
+        var kafka = new KafkaBuilder(KAFKA)
+                .editMetadata()
+                .withNamespace("my-kafka")
+                .endMetadata()
+                .build();
+        EventSourceContext<KafkaService> context = MapperTestSupport.mockContextContaining(service);
         var mapper = new StrimziKafkaSecondaryToKafkaServicePrimaryMapper(context);
 
         // When
-        Set<ResourceID> primaryResourceIDs = mapper.toPrimaryResourceIDs(KAFKA);
+        Set<ResourceID> primaryResourceIDs = mapper.toPrimaryResourceIDs(kafka);
 
         // Then
-        assertThat(primaryResourceIDs).containsExactly(ResourceID.fromResource(SERVICE));
+        assertThat(primaryResourceIDs).containsExactly(ResourceID.fromResource(service));
     }
+
 }